Complete joystick lib

+++ lib/joystick/src/joystick.h
@@ -11,7 +11,7 @@ public:
     float xmin;
     float ymin;
     joystick(float _x, float _y, float _xmax, float _ymax, float _xmin, float _ymin);
-    int vel;
+    float vel;
     float phi;
 };
 #endif
